+++ .dotnet/src/Custom/Chat/ChatCompletion.cs
@@ -7,8 +7,8 @@ namespace OpenAI.Chat;
 [CodeGenModel("CreateChatCompletionResponse")]
 public partial class ChatCompletion
 {
-    private IReadOnlyList<ChatTokenLogProbabilityInfo> _contentTokenLogProbabilities;
-    private IReadOnlyList<ChatTokenLogProbabilityInfo> _refusalTokenLogProbabilities;
+    private IReadOnlyList<ChatTokenLogProbabilityDetails> _contentTokenLogProbabilities;
+    private IReadOnlyList<ChatTokenLogProbabilityDetails> _refusalTokenLogProbabilities;
 
     // CUSTOM: Made private. This property does not add value in the context of a strongly-typed class.
     /// <summary> The object type, which is always `chat.completion`. </summary>
@@ -38,14 +38,14 @@ public partial class ChatCompletion
     /// <summary>
     /// Log probability information.
     /// </summary>
-    public IReadOnlyList<ChatTokenLogProbabilityInfo> ContentTokenLogProbabilities => (Choices[0].Logprobs != null)
+    public IReadOnlyList<ChatTokenLogProbabilityDetails> ContentTokenLogProbabilities => (Choices[0].Logprobs != null)
         ? Choices[0].Logprobs.Content
-        : _contentTokenLogProbabilities ??= new ChangeTrackingList<ChatTokenLogProbabilityInfo>();
+        : _contentTokenLogProbabilities ??= new ChangeTrackingList<ChatTokenLogProbabilityDetails>();
 
     // CUSTOM: Flattened refusal logprobs property.
-    public IReadOnlyList<ChatTokenLogProbabilityInfo> RefusalTokenLogProbabilities => (Choices[0]?.Logprobs != null)
+    public IReadOnlyList<ChatTokenLogProbabilityDetails> RefusalTokenLogProbabilities => (Choices[0]?.Logprobs != null)
         ? Choices[0].Logprobs.Refusal
-        : _refusalTokenLogProbabilities ??= new ChangeTrackingList<ChatTokenLogProbabilityInfo>();
+        : _refusalTokenLogProbabilities ??= new ChangeTrackingList<ChatTokenLogProbabilityDetails>();
 
     // CUSTOM: Flattened choice message property.
     /// <summary>