+++ .dotnet/src/Custom/Chat/ChatTokenLogProbabilityDetails.cs
@@ -1,5 +1,6 @@
 namespace OpenAI.Chat;
 
+using System;
 using System.Collections.Generic;
 
 /// <summary>
@@ -7,21 +8,23 @@ using System.Collections.Generic;
 /// <see cref="ChatCompletionOptions.IncludeLogProbabilities"/>.
 /// </summary>
 [CodeGenModel("ChatCompletionTokenLogprob")]
-public partial class ChatTokenLogProbabilityInfo
+public partial class ChatTokenLogProbabilityDetails
 {
     // CUSTOM: Renamed.
     /// <summary> The log probability of this token, if it is within the top 20 most likely tokens. Otherwise, the value `-9999.0` is used to signify that the token is very unlikely. </summary>
     [CodeGenMember("Logprob")]
     public float LogProbability { get; }
 
-    // CUSTOM: Renamed.
+    // CUSTOM:
+    // - Renamed.
+    // - Changed type from IReadOnlyList<int> to ReadOnlyMemory<byte>?.
     /// <summary>
     /// A list of integers representing the UTF-8 bytes representation of the token. Useful in instances where
     /// characters are represented by multiple tokens and their byte representations must be combined to generate
     /// the correct text representation. Can be null if there is no bytes representation for the token.
     /// </summary>
     [CodeGenMember("Bytes")]
-    public IReadOnlyList<int> Utf8ByteValues { get; }
+    public ReadOnlyMemory<byte>? Utf8Bytes { get; }
 
     // CUSTOM: Renamed.
     /// <summary>
@@ -30,5 +33,5 @@ public partial class ChatTokenLogProbabilityInfo
     /// <see cref="ChatCompletionOptions.TopLogProbabilityCount"/>.
     /// </summary>
     [CodeGenMember("TopLogprobs")]
-    public IReadOnlyList<ChatTokenTopLogProbabilityInfo> TopLogProbabilities { get; }
+    public IReadOnlyList<ChatTokenTopLogProbabilityDetails> TopLogProbabilities { get; }
 }