+++ .dotnet/src/Custom/Chat/ChatTokenTopLogProbabilityDetails.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace OpenAI.Chat;
@@ -8,19 +9,21 @@ namespace OpenAI.Chat;
 /// <see cref="ChatCompletionOptions.TopLogProbabilityCount"/>.
 /// </summary>
 [CodeGenModel("ChatCompletionTokenLogprobTopLogprob")]
-public partial class ChatTokenTopLogProbabilityInfo
+public partial class ChatTokenTopLogProbabilityDetails
 {
     // CUSTOM: Renamed.
     /// <summary> The log probability of this token, if it is within the top 20 most likely tokens. Otherwise, the value `-9999.0` is used to signify that the token is very unlikely. </summary>
     [CodeGenMember("Logprob")]
     public float LogProbability { get; }
 
-    // CUSTOM: Renamed.
+    // CUSTOM:
+    // - Renamed.
+    // - Changed type from IReadOnlyList<int> to ReadOnlyMemory<byte>?.
     /// <summary>
     /// A list of integers representing the UTF-8 bytes representation of the token. Useful in instances where
     /// characters are represented by multiple tokens and their byte representations must be combined to generate
     /// the correct text representation. Can be null if there is no bytes representation for the token.
     /// </summary>
     [CodeGenMember("Bytes")]
-    public IReadOnlyList<int> Utf8ByteValues { get; }
+    public ReadOnlyMemory<byte>? Utf8Bytes { get; }
 }